Remove tests for Symfony 3.0
Symfony 3.0 is no longer maintained and no longer compatible with twig 2
or 1.27+. There is no easy way to keep testing against it.

This only has effect on the tests, the bundle itself still supports
Symfony 3.0 (if it's installed with the correct Twig version, but that's a
bug of Symfony 3.0 and not of this bundle).