[BUGFIX] Prevent null-check on get_class occurrence
Since PHP's get_class() has the unlucky ability to pass null
(resp. nothing) as argument and considers the calling class
in this case which leads to side effects; and PHP 7.2 doesn't
allow null values at all anymore; this change fixes a case
where this might happen as of the interface since a null
object can be passed.